+++ src/plugins/proxy/commands.js
@@ -1,14 +1,14 @@
-import * as handlers from './command-handlers';
+import * as commandHandlers from './command-handlers';
 
 export const setup = {
   description: 'Setup and start proxy',
-  handler: handlers.setup
+  handler: commandHandlers.setup
 };
 
 export const reconfigShared = {
   name: 'reconfig-shared',
   description: 'Reconfigure shared properties',
-  handler: handlers.reconfigShared
+  handler: commandHandlers.reconfigShared
 };
 
 export const logs = {
@@ -16,20 +16,29 @@ export const logs = {
   builder(yargs) {
     return yargs.strict(false);
   },
-  handler: handlers.logs
+  handler: commandHandlers.logs
+};
+
+export const leLogs = {
+  name: 'logs-le',
+  description: 'View logs for Let\'s Encrypt',
+  builder(yargs) {
+    return yargs.strict(false);
+  },
+  handler: commandHandlers.leLogs
 };
 
 export const envconfig = {
   description: 'Configure environment variables for proxy',
-  handler: handlers.envconfig
+  handler: commandHandlers.envconfig
 };
 
 export const start = {
   description: 'Start proxy and let\'s encrypt containers',
-  handler: handlers.start
+  handler: commandHandlers.start
 };
 
 export const stop = {
   description: 'Stop proxy',
-  handler: handlers.stop
+  handler: commandHandlers.stop
 };